requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML Dependencies
torch==2.1.0
transformers==4.35.0
sentencepiece==0.1.99
accelerate==0.24.1

# Database
pymongo==4.6.0
motor==3.3.2

# Utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
